huggingface-space/requirements.txt
torch
torchvision
transformers
gradio>=3.50.2
numpy
Pillow
huggingface_hub